+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - Added the new `x/subspaces` module ([#392](https://github.com/desmos-labs/desmos/issues/392))
 - Added the ability to paginate the relationships of a profile ([#467](https://github.com/desmos-labs/desmos/issues/467)) 
 - Added the ability to edit whether users can comment on a post or not ([#446](https://github.com/desmos-labs/desmos/issues/446))
-
+- Added the ability to paginate the registered reactions ([#471](https://github.com/desmos-labs/desmos/issues/471))
 
 # Version 0.16.3
 ## Changes